% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_utils.R
\name{assign_genetic_map}
\alias{assign_genetic_map}
\title{Assign or Interpolate Genetic Map Values}
\usage{
assign_genetic_map(snp_df, map_df, strict = FALSE)
}
\arguments{
\item{snp_df}{dataframe of snp coordinates. `snp_df` must contain (integer-valued) columns named `chr` and `pos`.}

\item{map_df}{dataframe of reference genetic map values. `map_df` must contain integer-valued columns named `chr` and `pos`,
as well as a numeric-valued column called `map`, which must be a _strictly_ sorted vector of cumulative genetic map values.}

\item{strict}{boolean indicating whether}
}
\value{
a modified `snp_df` with an additional column giving the interpolated genetic map values at each locus
}
\description{
Assign or Interpolate Genetic Map Values
}
