% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_utils.R
\name{assign_map}
\alias{assign_map}
\title{Assign or Interpolate Genetic Map Values}
\usage{
assign_map(snp_df, map_df)
}
\arguments{
\item{snp_df}{}

\item{map_df}{dataframe of LD blocks, if `NULL`, use precomputed (EUR) ldshrink LD blocks}

\item{assign_all}{whether to throw an error if a SNP cannot be assigned to a block, or to assign it to block `NA`}
}
\description{
Assign or Interpolate Genetic Map Values
}
