% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_utils.R
\name{assign_region}
\alias{assign_region}
\title{Assign SNPs to LD blocks}
\usage{
assign_region(break_chr, break_start, break_stop,
  break_id = seq_along(break_chr), snp_chr, snp_pos, assign_all = T)
}
\arguments{
\item{break_chr}{integer vector of chromosome assignment for each LD block}

\item{break_start}{integer vector of the start coordinate for each LD block}

\item{break_stop}{integer vector of the stop coordinate for each LD block}

\item{break_id}{integer vector with a unique id for each LD block}

\item{snp_chr}{dataframe of snp coordinates, must contain (integer-valued) columns named `chr` and `pos`.}

\item{snp_pos}{integer vector giving the coordinate of each SNP}

\item{assign_all}{whether to throw an error if a SNP cannot be assigned to a block, or to assign it to block `NA`}

\item{snp_chr}{integer vector giving the chromosome of every SNP}
}
\value{
modified `snp_df` dataframe with additional column `region_id` mapping snp to LD block
}
\description{
Assign SNPs to LD blocks
}
