% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_theta}
\alias{calc_theta}
\title{Calculate the constant theta, given `m`}
\usage{
calc_theta(m)
}
\arguments{
\item{m}{a number indicating the size of the panel used to create the genetic map
(if using `1000-genomes-genetic-maps` from europeans, this number is 85)}
}
\description{
Calculate the constant theta, given `m`
}
