% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_utils.R
\name{chunk_genome}
\alias{chunk_genome}
\title{Assign SNPs to LD blocks}
\usage{
chunk_genome(snp_df, n_chunks = NA, chunk_size = NA, min_size = 1)
}
\arguments{
\item{snp_df}{dataframe of snp coordinates, must contain (integer-valued) columns named `chr` and `pos`.}

\item{n_chunks}{number of (approximately) equally sized chunks to be returned. (cannot be combined with `chunk_size`)}

\item{chunk_size}{maximum size of each chunk (cannot be combined with `n_chunks`)}
}
\value{
modified `snp_df` dataframe with additional column `region_id` mapping snp to LD block
}
\description{
Assign SNPs to LD blocks
}
