% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDshrink.R
\name{estimate_LD}
\alias{estimate_LD}
\title{Calculate LD from a reference panel}
\usage{
estimate_LD(reference_panel, map, method = "ldshrink",
  method_ld_params = list(m = 85, Ne = 11490.672741, cutoff = 0.001),
  output = "dsCMatrix", store_anno = FALSE, progress = FALSE,
  isGenotype = TRUE, ...)
}
\arguments{
\item{reference_panel}{is an `n` by `p` matrix of reference-panel haplotypes or genotypes which will be used for LD calculation}

\item{map}{is a length `p` vector of cumulative genetic map values. `map` must be _strictly_ _sorted_}

\item{method}{method for computing LD,(defaults to the Wen-Stephens 2010 shrinkage-estimator)  a list of available methods may be found in the `Details` section below}

\item{method_ld_params}{a list of additional arguments to be passed to the LD method}

\item{output}{a length one character vector specifying the format for output.  defaults to a sparse,symmetric matrix.}

\item{store_anno}{a boolean indicating whether to store the result of the distance function, can only be used with `output="data.frame"`}

\item{progress}{a boolean indicating whether to show a progress bar}

\item{isGenotype}{a boolean indicating whether the data is genotype data (the alternative being that the data is haplotype data)}
}
\value{
panel_ld an estiamate of LD
}
\description{
`ldshrink` constructs an LD matrix from a reference panel.
}
\examples{
#load data
data(reference_genotype)
data(reference_map)
#obtain a symmetric sparse
sparse_LD  <- estimate_LD(reference_genotype,reference_map)
#obtain a dense matrix
dense_LD  <- estimate_LD(reference_genotype,reference_map,output="matrix")
}
