% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDshrink.R
\name{estimate_LDscores}
\alias{estimate_LDscores}
\title{Estimate LD scores from the LD matrix}
\usage{
estimate_LDscores(R, N)
}
\arguments{
\item{R}{a pxp LD matrix (as obtained from `estimate_LD`)}

\item{N}{a scalar representing the number of samples in the reference panel}
}
\value{
A length p vector with LD scores at each locus
}
\description{
Estimate LD scores from the LD matrix
}
\examples{
data(reference_genotype)
data(reference_map)
R <- estimate_LD(reference_panel=reference_genotype,map=reference_map)
L2 <- estimate_LDscores(R,nrow(reference_genotype))
}
