% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ld2df}
\alias{ld2df}
\title{'Melt' an LD matrix, dropping elements below a given r-square cutoff}
\usage{
ld2df(ldmat, rsid, r2cutoff = 0.01, stringsAsFactors = FALSE)
}
\description{
'Melt' an LD matrix, dropping elements below a given r-square cutoff
}
