% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldshrink-package.R
\docType{package}
\name{ldshrink-package}
\alias{ldshrink}
\alias{ldshrink-package}
\title{ldshrink: A one-stop R package for shrinkage estimation of linkage disequilibrium.}
\description{
Implements TBA.  These methods can be applied to TBA.
    Note that 'ldshrink' is a companion package to 'rssr'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stephenslab/ldshrink}
  \item Report bugs at \url{https://github.com/stephenslab/ldshrink/issues}
}

}
\author{
\strong{Maintainer}: Xiang Zhu \email{xiangzhu@uchicago.edu}

Authors:
\itemize{
  \item Matthew Stephens \email{mstephens@uchicago.edu}
  \item Xin He \email{xinhe@uchicago.edu}
  \item Nicholas Knoblauch \email{nwknoblauch@uchicago.edu}
}

}
