% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDshrink.R
\name{ldshrink}
\alias{ldshrink}
\title{Calculate ldshrink adjusted LD matrix}
\usage{
ldshrink(
  genotype_panel,
  map_data,
  m = 85,
  Ne = 11490.672741,
  cutoff = 0.001,
  isGeno = NA,
  cov_2_cor = TRUE,
  na.rm = TRUE,
  process_map = TRUE
)
}
\arguments{
\item{genotype_panel}{`n` by `p` genotype or haplotype numeric matrix from a contiguous region of the genome ()}

\item{map_data}{vector of cumulative genetic map values.  Must be sorted}
}
\value{
LD matrix
}
\description{
Calculate ldshrink adjusted LD matrix
}
