% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDshrink.R
\name{ldshrink_evd}
\alias{ldshrink_evd}
\title{Calculate LD and eigenvalue decomposition from a reference panel}
\usage{
ldshrink_evd(reference_panel, map = NULL, m = 85, Ne = 11490.672741,
  cutoff = 0.001, useldshrink = T, na.rm = F, ...)
}
\arguments{
\item{reference_panel}{reference panel haplotype}

\item{map}{refrence genetic map}

\item{m}{sample size for genetic map}

\item{Ne}{effective population size}

\item{cutoff}{correlation cutoff}

\item{useldshrink}{whether or not to use LD shrink}

\item{na.rm}{whether or not to remove missing values}

\item{...}{extra arguments to other functions(}
}
\description{
Calculate LD and eigenvalue decomposition from a reference panel
}
