% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{reference_genotype}
\alias{reference_genotype}
\title{Sample genotype data}
\format{A data matrix with genotype from 503 individuals(rows) at 312 loci(columns)
the `reference_genotype` dataset}
\usage{
reference_genotype
}
\description{
Sample genotype data
}
\keyword{datasets}
