% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{reference_map}
\alias{reference_map}
\title{Sample genetic map data}
\format{A vector with genetic map data at 312 loci}
\usage{
reference_map
}
\description{
Sample genetic map data
}
\keyword{datasets}
