% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_utils.R
\name{assign_snp_block}
\alias{assign_snp_block}
\title{Assign SNPs to LD blocks}
\usage{
assign_snp_block(snp_df, break_df = NULL, assign_all = T)
}
\arguments{
\item{snp_df}{}

\item{break_df}{dataframe of LD blocks, if `NULL`, use precomputed (EUR) ldshrink LD blocks}

\item{assign_all}{whether to throw an error if a SNP cannot be assigned to a block, or to assign it to block `NA`}
}
\description{
Assign SNPs to LD blocks
}
