% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_utils.R
\name{chunk_genome}
\alias{chunk_genome}
\title{Assign SNPs to LD blocks}
\usage{
chunk_genome(snp_df, n_chunks = NA, chunk_size = NA, min_size = 10)
}
\arguments{
\item{snp_df}{dataframe of snp coordinates, must contain columns named `chr` and `pos`.}

\item{break_df}{dataframe of LD blocks, if `NULL`, use precomputed (EUR) ldshrink LD blocks, must contain columns named `chr` `start`, `stop`, and `region_id`}

\item{assign_all}{whether to throw an error if a SNP cannot be assigned to a block, or to assign it to block `NA`}
}
\value{
modified `snp_df` dataframe with additional column `region_id` mapping snp to LD block
}
\description{
Assign SNPs to LD blocks
}
